add createTodoItem class and correct datetime to instant

+++ src/Todo.Shared/TodoItemPrivate.cs
@@ -5,6 +5,6 @@ namespace TodoApi.Shared.Models{
         public int Id {get;set;}
         public bool IsComplete {get;set;}
         public int UserId { get; set; }
-        public int CreatedDate { get; set; }
+        public Int64 CreatedDate { get; set; }
     }
 }
